refactor: update mobile test screen imports to use primitives barrel
- Replace all subpath imports (theme/, Icon/, Callout/, etc.) with barrel
- Remove Container references, replace with View from react-native
- Remove FlexCenter/FlexRow references (don't exist)
- Merge duplicate import lines into single barrel imports
